% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdata}
\alias{gdata}
\title{gdata}
\usage{
gdata(h, p, s, M, sigbeta, key, sizeA0, nk)
}
\arguments{
\item{h}{Measure the degree of similarity between the source model and the target model parameters in set A}

\item{p}{Parameter dimension}

\item{s}{The number of non-zero elements in the parameter}

\item{M}{The number of source models}

\item{sigbeta}{One of the initialization values of the parameter}

\item{key}{1 or 2 represents different data settings}

\item{sizeA0}{Number of source models in set A}

\item{nk}{The sample size of source models}
}
\value{
B Generated parameters of target model and source models
}
\description{
A function that generates parameters of source models and target model
}
\examples{
\dontrun{
h=2; p=20; s=11; M=10; sigbeta=0.4; key=1;  sizeA0=4; nk=50
B <- gdata(h,p,s,M,sigbeta,key,sizeA0,nk)
}
}
