% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLMMA.R
\name{plotwto1}
\alias{plotwto1}
\title{plotwto1}
\usage{
plotwto1(h, p, s, M, sigbeta, sigx)
}
\arguments{
\item{h}{Measure the degree of similarity between the source model and the target model parameters in set A}

\item{p}{Parameter dimension}

\item{s}{The The number of non-zero elements in the parameter}

\item{M}{The The number of source models}

\item{sigbeta}{One of the initialization values of the parameter}

\item{sigx}{The covariance matrix of X}
}
\value{
A graph about the trend of the sum of weights of potential auxiliary models as the sample size increases
}
\description{
A Graph which used to verify the theory of the method TLMMA.
}
\examples{
\dontrun{
h=2;p=20;s=14;M=10;sigbeta=0.4;sigx=diag(1,p);
plotwto1(h,p,s,M,sigbeta,sigx)
}
}
