% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLMMA.R
\name{plotwto2}
\alias{plotwto2}
\title{plotwto2}
\usage{
plotwto2(h, p, s, n0, nk, M, sigbeta, sigx, key)
}
\arguments{
\item{h}{Measure the degree of similarity between the source model and the target model parameters in set A}

\item{p}{Parameter dimension}

\item{s}{The The number of non-zero elements in the parameter}

\item{n0}{The sample size of target model}

\item{nk}{The sample size of source models}

\item{M}{The The number of source models}

\item{sigbeta}{One of the initialization values of the parameter}

\item{sigx}{The covariance matrix of X}

\item{key}{1=Calculation method result 2=Verify relevant theories}
}
\value{
A graph about MSE and PMSE of the TLMMA method
}
\description{
Graphs are used to verify the performance of the method TLMMA.
}
\examples{
\dontrun{
h=6;p=20;s=11;n0=100;nk=100;M=20;sigbeta=0.3;sigx=diag(1,p);key=1;
plotwto2(h,p,s,n0,nk,M,sigbeta,sigx,key)
}
}
