% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLMMA.R
\name{TLMMA}
\alias{TLMMA}
\title{TLMMA}
\usage{
TLMMA(h, p, s, n0, nk, M, sigbeta, sigx, key)
}
\arguments{
\item{h}{Measure the degree of similarity between the source model and the target model parameters in set A}

\item{p}{Parameter dimension}

\item{s}{The The number of non-zero elements in the parameter}

\item{n0}{The sample size of target model}

\item{nk}{The sample size of source models}

\item{M}{The The number of source models}

\item{sigbeta}{One of the initialization values of the parameter}

\item{sigx}{The covariance matrix of X}

\item{key}{1 or 2 represents different data settings}
}
\value{
MSE pMSE weight betahat
}
\description{
TLMMA is a method to improve the estimation of target model by using multi-source domain samples under the background of transfer learning.
}
\examples{
\dontrun{
p=20;  h=2; s=11;  n0=100;  nk=50;  M=10;  sigbeta=0.4; sigx=diag(1,p); key=1
TLMMA(h,p,s,n0,nk,M,sigbeta,sigx,key)
}
}
